Moved rgb-to-hex to util
no issue

- Moved the rgb to hex function into a util so we can use it later in
  other areas.
